Rewrote using net-ssh directly, and reworking cd().
Using net-ssh directly mainly has the advantage of allowing us to get command
output as it comes rather than as a whole.
It also gets us rid of net-ssh-session's behavior of putting gibberish in the
shell's history.

cd() now works by predending the command with "cd path && cd path2..." chains.
Previously, we tried to get the current directory, cd to the new one, execute
the block and cd back to the previous one. This was extremely unreliable, as
the pwd command output included the user-defined prompt, which is probably
impossible reliably remove from the path.